# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/plans": {
      "post": {
        "operationId": "PlansController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created."
          }
        },
        "summary": "Create a new plan (platform_admin)",
        "tags": [
          "Plans"
        ]
      },
      "get": {
        "operationId": "PlansController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Get all plans (platform_admin)",
        "tags": [
          "Plans"
        ]
      }
    },
    "/plans/{id}": {
      "get": {
        "operationId": "PlansController_findOne",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Get a plan by ID (platform_admin)",
        "tags": [
          "Plans"
        ]
      },
      "patch": {
        "operationId": "PlansController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Update a plan (platform_admin)",
        "tags": [
          "Plans"
        ]
      },
      "delete": {
        "operationId": "PlansController_remove",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Delete a plan (platform_admin)",
        "tags": [
          "Plans"
        ]
      }
    },
...
    "/platform-support-access": {
      "get": {
        "operationId": "PlatformSupportAccessController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "adminUserId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List support accesses",
        "tags": [
          "Platform Support Access"
        ]
      },
      "post": {
        "operationId": "PlatformSupportAccessController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlatformSupportAccessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Grant support access",
        "tags": [
          "Platform Support Access"
        ]
      }
    },
    "/platform-support-access/{id}": {
      "get": {
        "operationId": "PlatformSupportAccessController_findOne",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a support access",
        "tags": [
          "Platform Support Access"
        ]
      },
      "patch": {
        "operationId": "PlatformSupportAccessController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlatformSupportAccessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a support access",
        "tags": [
          "Platform Support Access"
        ]
      },
      "delete": {
        "operationId": "PlatformSupportAccessController_remove",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `platform_admin`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Revoke support access",
        "tags": [
          "Platform Support Access"
        ]
      }
    },
...
  },
  "info": {
    "title": "OfficeSync API",
    "description": "Backend API for the OfficeSync HR and PM Dashboard",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Role-Based-Access": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Enter your role to unlock the endpoints (e.g., platform_admin, superuser, hr_manager, team_leader, team_member)"
      }
    },
    "schemas": {
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Enterprise"
          },
          "maxBranches": {
            "type": "number",
            "example": 10
          },
          "maxUsers": {
            "type": "number",
            "example": 100
          },
          "maxActiveProcessTemplates": {
            "type": "number",
            "example": 50
          },
          "maxComplianceRules": {
            "type": "number",
            "example": 25
          },
          "auditLogRetentionDays": {
            "type": "number",
            "example": 365
          },
          "allowsIntegrations": {
            "type": "boolean",
            "example": true
          },
          "monthlyPrice": {
            "type": "number",
            "example": 99.99
          },
          "annualPrice": {
            "type": "number",
            "example": 999.99
          },
          "currency": {
            "type": "string",
            "example": "USD"
          }
        },
        "required": [
          "name",
          "auditLogRetentionDays",
          "allowsIntegrations"
        ]
      },
      "UpdatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Enterprise"
          },
          "maxBranches": {
            "type": "number",
            "example": 10
          },
          "maxUsers": {
            "type": "number",
            "example": 100
          },
          "maxActiveProcessTemplates": {
            "type": "number",
            "example": 50
          },
          "maxComplianceRules": {
            "type": "number",
            "example": 25
          },
          "auditLogRetentionDays": {
            "type": "number",
            "example": 365
          },
          "allowsIntegrations": {
            "type": "boolean",
            "example": true
          },
          "monthlyPrice": {
            "type": "number",
            "example": 99.99
          },
          "annualPrice": {
            "type": "number",
            "example": 999.99
          },
          "currency": {
            "type": "string",
            "example": "USD"
          }
        }
      },
...
      "CreatePlatformSupportAccessDto": {
        "type": "object",
        "properties": {
          "adminUserId": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-09-01T00:00:00Z"
          }
        },
        "required": [
          "adminUserId",
          "companyId",
          "reason",
          "expiresAt"
        ]
      },
      "UpdatePlatformSupportAccessDto": {
        "type": "object",
        "properties": {
          "adminUserId": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-09-01T00:00:00Z"
          },
          "actionLog": {
            "description": "Audit log of actions taken during support session",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "Role-Based-Access": []
    }
  ]
}